plinkAPI/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["smzalam <alam.syedmz@gmail.com>"]
readme = "README.md"
packages = [{include = "server", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.104.1"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pandas = "^2.1.3"
numpy = "^1.26.2"
pydantic = "^2.5.2"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
pytest = "^7.4.3"
psycopg = "^3.1.13"
psycopg-pool = "^3.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
